requirements.txt
attrs==23.1.0
beautifulsoup4==4.12.2
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
et-xmlfile==1.1.0
exceptiongroup==1.1.3
google-auth==2.23.1
google-auth-oauthlib==1.1.0
greenlet==2.0.2
gspread==5.11.2
h11==0.14.0
idna==3.4
oauthlib==3.2.2
openpyxl==3.1.2
outcome==1.2.0
packaging==23.1
Pillow==10.0.1
playwright==1.38.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyee==9.0.4
PySocks==1.7.1
pytesseract==0.3.10
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
selenium==4.13.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
trio==0.22.2
trio-websocket==0.11.1
typing-extensions==4.8.0
urllib3==2.0.5
wsproto==1.2.0
